# devpick-api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trace API",
    "description": "Trace 백엔드 REST API — 개발자 성장형 통합 플랫폼\n\n**Base URL**: `https://api.devpick.kr/v1`\n\n**인증**: `Authorization: Bearer {access_token}` (JWT)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 개발 서버"
    },
    {
      "url": "https://api.devpick.kr/v1",
      "description": "프로덕션 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {},
          "error": {
            "type": "object",
            "nullable": true,
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "detail": {}
            }
          }
        }
      },
      "Job": {
        "type": "string",
        "enum": ["FRONTEND", "BACKEND", "FULLSTACK"]
      },
      "Level": {
        "type": "string",
        "enum": ["BEGINNER", "JUNIOR", "MIDDLE", "SENIOR"]
      },
...
      "AnswerListResponse": {
        "type": "object",
        "properties": {
          "answers": { "type": "array", "items": { "$ref": "#/components/schemas/AnswerWithCommentsResponse" } }
        }
      },
      "AnswerWithCommentsResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "content": { "type": "string" },
          "authorId": { "type": "string", "format": "uuid" },
          "authorNickname": { "type": "string" },
          "authorJob": { "$ref": "#/components/schemas/Job" },
          "authorLevel": { "$ref": "#/components/schemas/Level" },
          "authorProfileImage": { "type": "string", "nullable": true },
          "isAdopted": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "comments": { "type": "array", "items": { "$ref": "#/components/schemas/CommentResponse" } }
        }
      },
...
      "AnswerCreateRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": { "type": "string", "example": "Fetch Join을 사용하면..." }
        }
      },
      "AnswerUpdateRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": { "type": "string" }
        }
      },
      "AnswerResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "postId": { "type": "string", "format": "uuid" },
          "content": { "type": "string" },
          "isAdopted": { "type": "boolean" },
          "authorId": { "type": "string", "format": "uuid" },
          "authorNickname": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "answerId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "nickname": { "type": "string" },
          "content": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "AUTH_001" },
              "message": { "type": "string" },
              "detail": {}
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 필요",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": { "code": "AUTH_001", "message": "로그인이 필요합니다" } }
          }
        }
      },
      "NotFound": {
        "description": "리소스 없음",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/posts/{postId}/answers": {
      "get": {
        "tags": ["Community"],
        "summary": "답변 및 댓글 통합 조회",
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/AnswerListResponse" } } }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "tags": ["Community"],
        "summary": "답변 작성",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnswerCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/AnswerResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/posts/{postId}/answers/{answerId}": {
      "put": {
        "tags": ["Community"],
        "summary": "답변 수정",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "answerId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnswerUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/AnswerResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "작성자 아님" }
        }
      },
      "delete": {
        "tags": ["Community"],
        "summary": "답변 삭제",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "answerId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "삭제 성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "작성자 아님" }
        }
      }
    },
    "/posts/{postId}/answers/{answerId}/adopt": {
      "post": {
        "tags": ["Community"],
        "summary": "답변 채택",
        "description": "비관적 락(@Lock) 적용 — 동시 채택 방지. 게시글 작성자만 가능.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "answerId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "채택 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    { "properties": { "data": { "$ref": "#/components/schemas/AnswerResponse" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "게시글 작성자 아님" },
          "409": { "description": "이미 채택된 답변 있음" }
        }
      }
    },
    "/posts/{postId}/answers/{answerId}/like": {
      "post": {
        "tags": ["Community"],
        "summary": "답변 좋아요",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "answerId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "201": { "description": "성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "description": "이미 좋아요함" }
        }
      },
      "delete": {
        "tags": ["Community"],
        "summary": "답변 좋아요 취소",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "answerId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "취소 성공" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "좋아요 없음" }
        }
      }
    },
...
  },
...
}